habab.ru/requirements.txt
Flask
python-dotenv
requests
PyYAML
markdown
beautifulsoup4
markitdown[all] # Добавляем markitdown со всеми зависимостями
